- Добавил пул объектов - Генерация мира по сети - Пофиксил переход камеры по комнатам - Сделал стрельбу с пулом пуль - Добавил настройки пули (Bullet Data) - И еще куча мелких вещей

+++ Cyber Infection/Assets/CyberInfection/Scripts/GameMechanics/IAlive.cs
@@ -2,8 +2,8 @@
 {
 	public interface IAlive
 	{
-		int health { get; set; }
-		void GetDamage(int damageAmount);
+		float health { get; set; }
+		void GetDamage(float damageAmount);
 		void Die();
 	}
 }
